Add explicit types

+++ python/test/unit/fem/test_mixed_mesh_dofmap.py
@@ -25,12 +25,18 @@ def test_dofmap_mixed_topology():
     tri = [0, 1, 4, 0, 3, 4]
     quad = [1, 4, 2, 5]
     # cells with global indexing
-    cells = [np.array([t + 3 * rank for t in tri]), np.array([q + 3 * rank for q in quad])]
-    orig_index = [np.array([3 * rank, 1 + 3 * rank]), np.array([2 + 3 * rank])]
+    cells = [
+        np.array([t + 3 * rank for t in tri], dtype=np.int64),
+        np.array([q + 3 * rank for q in quad], dtype=np.int64),
+    ]
+    orig_index = [
+        np.array([3 * rank, 1 + 3 * rank], dtype=np.int64),
+        np.array([2 + 3 * rank], dtype=np.int64),
+    ]
     # No ghosting
-    ghost_owners = [np.array([]), np.array([])]
+    ghost_owners = [np.array([], dtype=np.int32), np.array([], dtype=np.int32)]
     # All vertices are on boundary
-    boundary_vertices = np.array([3 * rank + i for i in range(6)])
+    boundary_vertices = np.array([3 * rank + i for i in range(6)], dtype=np.int64)
 
     topology = create_topology(
         MPI.COMM_WORLD,
@@ -83,13 +89,13 @@ def test_dofmap_mixed_topology():
 
 def test_dofmap_prism_mesh():
     # Prism mesh
-    cells = [np.array([0, 1, 2, 3, 4, 5])]
+    cells = [np.array([0, 1, 2, 3, 4, 5], dtype=np.int64)]
     # cells with global indexing
-    orig_index = [np.array([0, 1, 2, 3, 4, 5])]
+    orig_index = [np.array([0, 1, 2, 3, 4, 5], dtype=np.int64)]
     # No ghosting
-    ghost_owners = [np.array([])]
+    ghost_owners = [np.array([], dtype=np.int32)]
     # All vertices are on boundary
-    boundary_vertices = np.array([0, 1, 2, 3, 4, 5])
+    boundary_vertices = np.array([0, 1, 2, 3, 4, 5], dtype=np.int64)
 
     topology = Topology(
         create_topology(
